refactor: break up CLIApplication into separate files (#1078)
* refactor: extract commands into separate files

* refactor: move helper functions into UserSessionScope extension

* refactor: move functions inside ConsoleCommand

* refactor: fix compiler warnings

* refactor: delete unused class
